pyproject.toml
[tool.poetry]
name = "soccer-video-analytics"
version = "0.1.0"
description = ""
authors = ["Diego Marvid <dmarvid@tryolabs.com>", "Diego Fernandez <dfernandez@tryolabs.com"]
packages = [
    { include = "soccer" },
    { include = "inference" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
norfair = {git = "https://github.com/tryolabs/norfair.git", rev = "master"}
torch = "^2.7.1"
opencv-python = "^4.6.0.66"
numpy = "^1.26.4"
Pillow = "^11.2.1"
matplotlib = "^3.10.3"
pandas = "^2.3.0"
torchvision = "^0.22.1"
PyYAML = "^6.0.2"
tqdm = "^4.64.1"
seaborn = "^0.12.0"
ipython = "^8.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
pytest-mock = "^3.12.0"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
mypy_path = "stubs"

[[tool.mypy.overrides]]
module = "norfair.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pyright]
reportMissingTypeStubs = false

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools>=42", "wheel"]
build-backend = "poetry.core.masonry.api"
